- clean up public classes - refactor per code climate warnings
